# Push takeString logic to the lexers for better performance